Enable --init for podman container
We had some issues with signals/killing child processes, which was fixed by running sshd in a loop. However, using /bin/sh as PID1 is not ideal, as it might not reap dead processes or handle signal forwarding, so it's better to use a small init process, as we're running full-fledged long-term environment in a container, and not simple well-known service from the image itself.